+++ src/main/java/ru/alfabouh/jgems3d/engine/render/opengl/frustum/FrustumCulling.java
@@ -1,29 +1,29 @@
 package ru.alfabouh.jgems3d.engine.render.opengl.frustum;
 
 import org.joml.FrustumIntersection;
-import org.joml.Matrix4d;
 import org.joml.Matrix4f;
-import org.joml.Vector4d;
+import org.joml.Matrix4f;
+import org.joml.Vector4f;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class FrustumCulling {
-    private final List<Vector4d> planes;
+    private final List<Vector4f> planes;
     private final FrustumIntersection frustumIntersection;
-    private Matrix4d projectionViewMatrix;
+    private Matrix4f projectionViewMatrix;
 
     public FrustumCulling() {
-        this.projectionViewMatrix = new Matrix4d();
+        this.projectionViewMatrix = new Matrix4f();
         this.planes = new ArrayList<>();
         this.frustumIntersection = new FrustumIntersection();
         for (int i = 0; i < 6; i++) {
-            this.planes.add(i, new Vector4d());
+            this.planes.add(i, new Vector4f());
         }
     }
 
-    public void refreshFrustumCullingState(Matrix4d projection, Matrix4d view) {
-        this.projectionViewMatrix = new Matrix4d();
+    public void refreshFrustumCullingState(Matrix4f projection, Matrix4f view) {
+        this.projectionViewMatrix = new Matrix4f();
         this.projectionViewMatrix.mul(projection);
         this.projectionViewMatrix.mul(view);
         for (int i = 0; i < 6; i++) {
