+++ src/main/java/ru/alfabouh/jgems3d/engine/render/opengl/scene/fabric/render/RenderParticle.java
@@ -11,7 +11,7 @@ public class RenderParticle extends RenderWorldItem {
     }
 
     @Override
-    public void onRender(double partialTicks, SceneRenderBase sceneRenderBase, IRenderObject renderItem) {
+    public void onRender(float partialTicks, SceneRenderBase sceneRenderBase, IRenderObject renderItem) {
         PhysicsObject entityObject = (PhysicsObject) renderItem;
         if (entityObject.isHasModel()) {
             JGemsSceneUtils.renderSceneObject(entityObject, entityObject.getRenderData().getOverObjectMaterial());
