+++ src/main/java/ru/alfabouh/jgems3d/engine/render/transformation/CameraTransformation.java
@@ -1,20 +1,20 @@
 package ru.alfabouh.jgems3d.engine.render.transformation;
 
-import org.joml.Matrix4d;
+import org.joml.Matrix4f;
 import ru.alfabouh.jgems3d.engine.render.opengl.scene.world.camera.ICamera;
 
 public class CameraTransformation {
-    private final Matrix4d viewMatrix;
+    private final Matrix4f viewMatrix;
 
     public CameraTransformation() {
-        this.viewMatrix = new Matrix4d().identity();
+        this.viewMatrix = new Matrix4f().identity();
     }
 
     public void update(ICamera camera) {
         this.viewMatrix.set(Transformation.getViewMatrix(camera));
     }
 
-    public Matrix4d getViewMatrix() {
-        return new Matrix4d(this.viewMatrix);
+    public Matrix4f getViewMatrix() {
+        return new Matrix4f(this.viewMatrix);
     }
 }
